pyproject.toml
[project]
name = "secure-flower-store-backend"
version = "0.1.0"
description = "FastAPI backend for a secure flower store"
authors = ["Your Name <you@example.com>"]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "SQLAlchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.6.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.0",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["app"]